🚚 (#266): Fix formatting + move az-document-ai to it's own submodule
