--- backend/app.hopps.az-document-ai/src/main/java/app/hopps/model/TradePartyHelper.java
@@ -1,24 +1,0 @@
-package app.hopps.model;
-
-import app.hopps.commons.TradeParty;
-import com.azure.ai.documentintelligence.models.AddressValue;
-
-public class TradePartyHelper {
-    private TradePartyHelper() {
-        // only call the static method
-    }
-
-    public static TradeParty fromAzure(AddressValue value) {
-        return new TradeParty(
-                null,
-                value.getCountryRegion(),
-                value.getPostalCode(),
-                value.getState(),
-                value.getCity(),
-                value.getRoad() + " " + value.getHouseNumber(),
-                value.getUnit(),
-                null,
-                null,
-                null);
-    }
-}
