+++ backend/app.hopps.az-document-ai/src/test/java/app/hopps/az/document/ai/AzureAiServiceMockedTest.java
@@ -1,4 +1,4 @@
-package app.hopps;
+package app.hopps.az.document.ai;
 
 import app.hopps.commons.DocumentData;
 import app.hopps.commons.DocumentType;
