+++ backend/app.hopps.az-document-ai/src/test/java/app/hopps/az/document/ai/ScanDocumentResourceTest.java
@@ -1,4 +1,4 @@
-package app.hopps;
+package app.hopps.az.document.ai;
 
 import app.hopps.commons.Address;
 import app.hopps.commons.InvoiceData;
